Add Rcode that writes a household_coordinates file
A household coordinates file is essential to creating the final population file. Also, this file was originally created on a different repository, but like most of the commits that I will do on this branch, it is brought in here as a way to organize the creation of a Salt Lake City scenario in MatSim.